% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{filter_date}
\alias{filter_date}
\title{Filter data frame dates}
\usage{
filter_date(df, date_f, r6_data)
}
\arguments{
\item{df}{data frame to filter}

\item{date_f}{field in data frame that has the date}

\item{r6_data}{R6 object of class app_data}
}
\description{
Filters dates in data frame using the start and end date from the
R6 app_data class object (noting that these are set via the inputs)
}
