% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{filter_ref_date}
\alias{filter_ref_date}
\title{Filter dates from Date Range & Period UI input}
\usage{
filter_ref_date(df, period, start, end)
}
\arguments{
\item{df}{data frame with data to filter}

\item{period}{period input from UI}

\item{start}{start date}

\item{end}{end date}
}
\description{
Note: currently filters using a reference data frame, needs testing against
approach that does it on the fly.
}
